KIT-230 refactor requiredNonEmptyString
https://coveord.atlassian.net/browse/KIT-230
