+++ packages/headless/src/features/context/context-actions.ts
@@ -1,9 +1,12 @@
 import {createAction} from '@reduxjs/toolkit';
 import {Context, ContextValue} from './context-state';
-import {validatePayload} from '../../utils/validate-payload';
-import {StringValue, isString, ArrayValue} from '@coveo/bueno';
+import {
+  validatePayload,
+  requiredNonEmptyString,
+} from '../../utils/validate-payload';
+import {isString, ArrayValue} from '@coveo/bueno';
 
-const nonEmptyString = new StringValue({required: true, emptyAllowed: false});
+const nonEmptyString = requiredNonEmptyString;
 const nonEmptyArray = new ArrayValue({
   each: nonEmptyString,
   required: true,
