+++ packages/headless/src/features/facets/facet-search-set/category/category-facet-search-actions.ts
@@ -1,21 +1,22 @@
 import {createAction} from '@reduxjs/toolkit';
 import {FacetSearchOptions} from '../facet-search-request-options';
 import {CategoryFacetSearchResult} from '../../../../api/search/facet-search/category-facet-search/category-facet-search-response';
-import {validatePayload} from '../../../../utils/validate-payload';
 import {
-  facetIdDefinition,
+  validatePayload,
   requiredNonEmptyString,
-} from '../../generic/facet-actions-validation';
-import {NumberValue, RecordValue, ArrayValue, StringValue} from '@coveo/bueno';
+  requiredEmptyAllowedString,
+} from '../../../../utils/validate-payload';
+import {facetIdDefinition} from '../../generic/facet-actions-validation';
+import {NumberValue, RecordValue, ArrayValue} from '@coveo/bueno';
 import {facetSearchOptionsDefinition} from '../generic/generic-facet-search-validate-payload';
 
 const categoryFacetSearchResultDefinition = {
   path: new ArrayValue({
     required: true,
-    each: new StringValue({required: true, emptyAllowed: false}),
+    each: requiredNonEmptyString,
   }),
-  displayValue: requiredNonEmptyString,
-  rawValue: requiredNonEmptyString,
+  displayValue: requiredEmptyAllowedString,
+  rawValue: requiredEmptyAllowedString,
   count: new NumberValue({required: true, min: 0}),
 };
