+++ packages/headless/src/features/facets/generic/facet-actions-validation.ts
@@ -1,11 +1,3 @@
-import {StringValue} from '@coveo/bueno';
+import {requiredNonEmptyString} from '../../../utils/validate-payload';
 
-export const requiredNonEmptyString = new StringValue({
-  required: true,
-  emptyAllowed: true,
-});
-
-export const facetIdDefinition = new StringValue({
-  required: true,
-  emptyAllowed: true,
-});
+export const facetIdDefinition = requiredNonEmptyString;
